+++ lambdas/data-migration2/tests/test-granulesAndFiles.js
@@ -116,7 +116,7 @@ test.beforeEach(async (t) => {
   );
   t.context.collectionPgModel = collectionPgModel;
   t.context.testCollection = testCollection;
-  t.context.collectionCumulusId = collectionResponse[0];
+  t.context.collectionCumulusId = collectionResponse[0].cumulus_id;
 
   const executionPgModel = new ExecutionPgModel();
   t.context.executionUrl = cryptoRandomString({ length: 5 });