+++ packages/api/lib/publishSnsMessageUtils.js
@@ -12,6 +12,32 @@ const publishExecutionSnsMessage = async (record) => {
   await publishSnsMessage(topicArn, record);
 };
 
+const constructSnsMessage = (record, event) => {
+  switch (event) {
+  case 'Create':
+  case 'Update':
+    return { event, record };
+  case 'Delete': return {
+    event,
+    record: {
+      name: record.name,
+      version: record.version,
+    },
+    deletedAt: Date.now(),
+  };
+  default: return {};
+  }
+};
+
+const publishCollectionSnsMessage = async (record, event) => {
+  const topicArn = envUtils.getRequiredEnvVar('collection_sns_topic_arn', process.env);
+  const messageToPublish = constructSnsMessage(record, event);
+
+  logger.info(`About to publish SNS message ${JSON.stringify(messageToPublish)} for collection to topic ARN ${topicArn}`);
+  await publishSnsMessage(topicArn, messageToPublish);
+};
+
 module.exports = {
+  publishCollectionSnsMessage,
   publishExecutionSnsMessage,
 };