+++ packages/api/tests/lib/test-publishSnsMessageUtils.js
@@ -1,37 +1,29 @@
 'use strict';
 
 const test = require('ava');
+const sinon = require('sinon');
 const cryptoRandomString = require('crypto-random-string');
 
 const { sns, sqs } = require('@cumulus/aws-client/services');
 
 const {
   publishExecutionSnsMessage,
+  publishCollectionSnsMessage,
 } = require('../../lib/publishSnsMessageUtils');
 
 const {
   fakeExecutionFactoryV2,
+  fakeCollectionFactory,
 } = require('../../lib/testUtils');
 
-test.serial('publishExecutionSnsMessage() does not publish an SNS message if execution_sns_topic_arn is undefined', async (t) => {
-  const newExecution = fakeExecutionFactoryV2({
-    arn: cryptoRandomString({ length: 5 }),
-    status: 'completed',
-    name: 'test_execution',
-  });
-  await t.throwsAsync(
-    publishExecutionSnsMessage(newExecution),
-    { message: 'The execution_sns_topic_arn environment variable must be set' }
-  );
-});
-
-test.serial('publishExecutionSnsMessage() publishes an SNS message', async (t) => {
+test.before(async (t) => {
   const topicName = cryptoRandomString({ length: 10 });
   const { TopicArn } = await sns().createTopic({ Name: topicName }).promise();
-  process.env.execution_sns_topic_arn = TopicArn;
+  t.context.TopicArn = TopicArn;
 
   const QueueName = cryptoRandomString({ length: 10 });
   const { QueueUrl } = await sqs().createQueue({ QueueName }).promise();
+  t.context.QueueUrl = QueueUrl;
   const getQueueAttributesResponse = await sqs().getQueueAttributes({
     QueueUrl,
     AttributeNames: ['QueueArn'],
@@ -49,6 +41,30 @@ test.serial('publishExecutionSnsMessage() publishes an SNS message', async (t) =
     Token: SubscriptionArn,
   }).promise();
 
+});
+
+test.after.always(async (t) => {
+  const { QueueUrl, TopicArn } = t.context;
+
+  await Promise.all([
+    sqs().deleteQueue({ QueueUrl }).promise(),
+    sns().deleteTopic({ TopicArn }).promise(),
+  ]);
+});
+
+test.serial('publishExecutionSnsMessage() does not publish an SNS message if execution_sns_topic_arn is undefined', async (t) => {
+  const newExecution = fakeExecutionFactoryV2({
+    arn: cryptoRandomString({ length: 5 }),
+    status: 'completed',
+    name: 'test_execution',
+  });
+  await t.throwsAsync(
+    publishExecutionSnsMessage(newExecution),
+    { message: 'The execution_sns_topic_arn environment variable must be set' }
+  );
+});
+
+test.serial('publishExecutionSnsMessage() publishes an SNS message', async (t) => {
   const executionArn = cryptoRandomString({ length: 10 });
   const newExecution = fakeExecutionFactoryV2({
     arn: executionArn,
@@ -57,7 +73,10 @@ test.serial('publishExecutionSnsMessage() publishes an SNS message', async (t) =
   });
   await publishExecutionSnsMessage(newExecution);
 
-  const { Messages } = await sqs().receiveMessage({ QueueUrl, WaitTimeSeconds: 10 }).promise();
+  const { Messages } = await sqs().receiveMessage({
+    QueueUrl: t.context.QueueUrl,
+    WaitTimeSeconds: 10,
+  }).promise();
 
   t.is(Messages.length, 1);
 
@@ -66,4 +85,82 @@ test.serial('publishExecutionSnsMessage() publishes an SNS message', async (t) =
 
   t.deepEqual(executionRecord.arn, executionArn);
   t.deepEqual(executionRecord.status, newExecution.status);
+});
+
+test.serial('publishCollectionSnsMessage() does not publish an SNS message if collection_sns_topic_arn is undefined', async (t) => {
+  const newCollection = fakeCollectionFactory();
+
+  t.teardown(() => {
+    process.env.collection_sns_topic_arn = t.context.TopicArn;
+  });
+
+  await t.throwsAsync(
+    publishCollectionSnsMessage(newCollection),
+    { message: 'The collection_sns_topic_arn environment variable must be set' }
+  );
+});
+
+test.serial('publishCollectionSnsMessage() publishes an SNS message for the event type Create', async (t) => {
+  process.env.collection_sns_topic_arn = t.context.TopicArn;
+  const collectionName = cryptoRandomString({ length: 10 });
+  const newCollection = fakeCollectionFactory({ name: collectionName });
+  await publishCollectionSnsMessage(newCollection, 'Create');
+
+  const { Messages } = await sqs().receiveMessage({
+    QueueUrl: t.context.QueueUrl,
+    WaitTimeSeconds: 10,
+  }).promise();
+
+  t.is(Messages.length, 1);
+
+  const snsMessage = JSON.parse(Messages[0].Body);
+  const message = JSON.parse(snsMessage.Message);
+
+  t.deepEqual(message.record, newCollection);
+  t.is(message.event, 'Create');
+});
+
+test.serial('publishCollectionSnsMessage() publishes an SNS message for the event type Update', async (t) => {
+  process.env.collection_sns_topic_arn = t.context.TopicArn;
+  const collectionName = cryptoRandomString({ length: 10 });
+  const newCollection = fakeCollectionFactory({ name: collectionName });
+  await publishCollectionSnsMessage(newCollection, 'Update');
+
+  const { Messages } = await sqs().receiveMessage({
+    QueueUrl: t.context.QueueUrl,
+    WaitTimeSeconds: 10,
+  }).promise();
+  t.is(Messages.length, 1);
+
+  const snsMessage = JSON.parse(Messages[0].Body);
+  const message = JSON.parse(snsMessage.Message);
+
+  t.deepEqual(message.record, newCollection);
+  t.is(message.event, 'Update');
+});
+
+test.serial('publishCollectionSnsMessage() publishes an SNS message for the event type Delete', async (t) => {
+  process.env.collection_sns_topic_arn = t.context.TopicArn;
+  const deletedAt = Date.now();
+  const stub = sinon.stub(Date, 'now').returns(deletedAt);
+  t.teardown(() => {
+    stub.restore();
+  });
+
+  const collectionName = cryptoRandomString({ length: 10 });
+  const newCollection = fakeCollectionFactory({ name: collectionName });
+  await publishCollectionSnsMessage(newCollection, 'Delete');
+
+  const { Messages } = await sqs().receiveMessage({
+    QueueUrl: t.context.QueueUrl,
+    WaitTimeSeconds: 10,
+  }).promise();
+  t.is(Messages.length, 1);
+
+  const snsMessage = JSON.parse(Messages[0].Body);
+  const message = JSON.parse(snsMessage.Message);
+
+  t.deepEqual(message.record, { name: newCollection.name, version: newCollection.version });
+  t.is(message.event, 'Delete');
+  t.is(message.deletedAt, deletedAt);
 });